+++ app/src/main/java/com/example/deivi/pedidosonline/EliminarMenu.java
@@ -75,7 +75,7 @@ public class EliminarMenu extends AppCompatActivity {
                         menus.setId(object.getString("_id"));
                         menus.setNombre(object.getString("nombre"));
                         menus.setDescripcion(object.getString("descripcion"));
-                        menus.setPrecio(object.getDouble("precio"));
+                        menus.setPrecio(object.getInt("precio"));
                         //menus.setFoto(object.getString("foto"));
                         list_data2.add(menus);
                     }
