+++ app/src/main/java/com/example/deivi/pedidosonline/Menu.java
@@ -98,7 +98,7 @@ public class Menu extends AppCompatActivity {
                         menus.setId(object.getString("_id"));
                         menus.setNombre(object.getString("nombre"));
                         menus.setDescripcion(object.getString("descripcion"));
-                        menus.setPrecio(object.getDouble("precio"));
+                        menus.setPrecio(object.getInt("precio"));
                         //menus.setFoto(object.getString("foto"));
                         list_data.add(menus);
                     }
