+++ app/src/main/java/com/example/deivi/pedidosonline/VerMenu.java
@@ -60,7 +60,7 @@ public class VerMenu extends AppCompatActivity {
                         menus.setId(object.getString("_id"));
                         menus.setNombre(object.getString("nombre"));
                         menus.setDescripcion(object.getString("descripcion"));
-                        menus.setPrecio(object.getDouble("precio"));
+                        menus.setPrecio(object.getInt("precio"));
                         //menus.setFoto(object.getString("foto"));
                         list_data.add(menus);
                     }
